py-grpc/requirements.txt
blinker==1.7.0
click==8.1.7
Flask==3.0.0
greenlet==3.0.2
grpcio==1.60.0
grpcio-tools==1.60.0
itsdangerous==2.1.2
Jinja2==3.1.2
MarkupSafe==2.1.3
protobuf==4.25.1
SQLAlchemy==2.0.23
typing_extensions==4.9.0
Werkzeug==3.0.1